refactor(api): move Accessibility enum from Challenge to CalibratedChallenge
