+++ api/src/certification/evaluation/domain/usecases/simulate-flash-assessment-scenario.js
@@ -1,6 +1,5 @@
 /**
- * @typedef {import ('../../../shared/domain/models/ComplementaryCertificationKeys.js').ComplementaryCertificationKeys} ComplementaryCertificationKeys
- * @typedef {import('./index.js').SharedChallengeRepository} SharedChallengeRepository
+ * @typedef {import('./index.js').CalibratedChallengeRepository} CalibratedChallengeRepository
  * @typedef {import('./index.js').VersionRepository} VersionRepository
  */
 
@@ -12,7 +11,7 @@ import { FlashAssessmentAlgorithm } from '../models/FlashAssessmentAlgorithm.js'
 /**
  * @param {Object} params
  * @param {number} params.stopAtChallenge - force scenario to stop at challenge before maximumAssessmentLength
- * @param {SharedChallengeRepository} params.sharedChallengeRepository
+ * @param {CalibratedChallengeRepository} params.calibratedChallengeRepository
  * @param {VersionRepository} params.versionRepository
  */
 export async function simulateFlashAssessmentScenario({
@@ -22,10 +21,10 @@ export async function simulateFlashAssessmentScenario({
   initialCapacity,
   variationPercent,
   flashAlgorithmService,
-  sharedChallengeRepository,
   accessibilityAdjustmentNeeded,
   stopAtChallenge,
   versionId,
+  calibratedChallengeRepository,
   versionRepository,
 }) {
   const version = await versionRepository.getById(versionId);
@@ -33,7 +32,7 @@ export async function simulateFlashAssessmentScenario({
   return _simulateCertificationScenario({
     locale,
     accessibilityAdjustmentNeeded,
-    challengeRepository: sharedChallengeRepository,
+    challengeRepository: calibratedChallengeRepository,
     flashAlgorithmService,
     pickChallenge,
     pickAnswerStatus,
@@ -46,7 +45,7 @@ export async function simulateFlashAssessmentScenario({
 
 /**
  * @param {Object} params
- * @param {SharedChallengeRepository} params.challengeRepository
+ * @param {CalibratedChallengeRepository} params.challengeRepository
  */
 async function _simulateCertificationScenario({
   pickChallenge,
