+++ api/tests/tooling/domain-builder/factory/build-calibrated-challenge.js
@@ -1,12 +1,15 @@
-import { CalibratedChallenge } from '../../../../src/certification/evaluation/domain/models/CalibratedChallenge.js';
+import {
+  Accessibility,
+  CalibratedChallenge,
+} from '../../../../src/certification/evaluation/domain/models/CalibratedChallenge.js';
 import { buildCalibratedChallengeSkill } from './build-calibrated-challenge-skill.js';
 
 export function buildCalibratedChallenge({
   id = 'recCHAL1',
   discriminant = 1,
   difficulty = 0,
-  blindnessCompatibility = 'OK',
-  colorBlindnessCompatibility = 'RAS',
+  blindnessCompatibility = Accessibility.OK,
+  colorBlindnessCompatibility = Accessibility.RAS,
   competenceId = 'competenceId',
   skill = buildCalibratedChallengeSkill(),
 } = {}) {
